added check if stocks are enough before order creation and items added/updated in cart

+++ src/main/java/com/harriet/shopiify/product/dto/ProductDTO.java
@@ -21,7 +21,7 @@ public class ProductDTO implements Serializable {
 
     private String units;
 
-    private BigInteger stock;
+    private Long stock;
 
     private Double price;
 }
